Accuracy Improvements
Discovered that accuracy issue is related to rounding based on bucket size, managed to reduce to 512 without issues and increases accuracy slightly, but need to figure out how to lower to ~128 without window breaking.

+++ FYP-DLL/AlgorithmicBeatMapping/AudioHandler.cpp
@@ -64,6 +64,7 @@ bool AudioHandler::performBeatMapping()
 	SpectralFluxAnalyser sfa;
 
 	std::vector<double> workSamples;
+	//workSamples = samples;
 
 	FFT fftHandle;
 	std::vector<double> workingSpectrum;
@@ -224,7 +225,7 @@ bool AudioHandler::performBeatMapping()
 		{
 			sendSpectrum.push_back(workingSpectrum[j]);
 		}
-		sfa.AnalyseSpectrum(sendSpectrum, (float)((1.0 / sampleRate) * i) * bucketSize);
+		sfa.AnalyseSpectrum(sendSpectrum, (1.0 / sampleRate) * i * bucketSize);
 		workingSpectrum.erase(workingSpectrum.begin(), workingSpectrum.begin() + bucketSize);
 	}
